feat: display bits for zone picker
This diff just gets the zones rendering as an unordered list in plain
html.

There's a bit of a shortfall here, aside from the styling. There's no
state management so selecting a zone is not yet possible.

There's also a fixme related to the "All" or "Unfiltered" state which
was previously represented at the top of the list via an "empty" list
item.

I'm not sure how I want to handle this use case in the Yew
version, but I'll revisit this after I get some state management going.