refactor: make Database a global const
Initially, I figured being able to pass in different values for the data
to `Database::new()` could be useful for unit tests, but I can get the same
from setting those private fields directly.

Also, I probably won't be writing tests for `Database`. This is a toy
project, after all.

Having the standard dataset available as a global should make access
from different components more convenient.

In the original JS version, all this data was set to keys in the redux
store, but I think this probably works just as well for the use case.