feat: collectible list
Build out the state management and DOM eventing needed to toggle the
checklist items, both in the document and in the store.

When (later) local storage is brought into the mix, we'll do a read at
the top of the component tree to populate the initial state of the
Store, then also update the storage via the reducer when it handles
`ToggleCollectible` actions.