feat: rough tabbed UI
It doesn't look like much, but this diff adds a couple of list items to
the top of the page. Clicking on them will control the visibility of the
collectible list and zone picker.

This is a tabbed UI in essence.

In the original React version, this was presented as a pair of icons
with a change in background color to indicate which was active.
Something similar should be possible once styling is underway.

Further refactors to tidy up are desirable, but for now this will do.