chore: add tailwind
Using the tailwindcss cli: `tailwindcss init` to produce a default
config.

Added bare-bones input.css to bring in the tailwind stuff, a glob to search
for class names under the rust source tree (where
our components are defined) in the config, and a `<link/>` in the
index.html so trunk can handle the style compilation.

Not yet sure if more configuration is needed to perform pruning of the
stylesheet etc.