feat: use scss for styling
Instead of trying to rebuild with tailwind, ths approach converts some
of the original less styles into scss (which is supported by the
bundler).

Since semantic-ui is not being used, there are some gaps/tweaks needed
to get us a finished design, but it looks like tailwind wouldn't help
much in this regard anyway.

Additional: the zone picker has been converted to be a plain select
element. Some additional styling will be needed to polish this all up (a
fixed position header might be nice), but I think it can wait.