feat: persist collected items in local storage
This diff will read from local storage on first render and use any
perviously persisted data (when set) to initialize the `Store` data.

When the store subsequently changes, we write the set of collected ids
back to the same local storage key.

This should cover the basic need for retaining checked items after the
browser tab is closed and reopened.