feat: order groups by active and uncollected, then name
This diff will float groups to the top when they contain any uncollected
items that belong to the selected (active) zone.

More could be done along these lines, such as sorting down any items
that are collected within each group, but for now I'll leave the
group-interior ordering as-is.

This diff brings us to feature-parity with the original React/Rocket
app.