chore: style fixups (#15)
* fix: set z-index on topbar

There was an issue, perhaps due to the opacity change on list items,
which could cause translucent items to render on top of the topbar
while scrolling.

Setting the z-index on the top bar seems to resolve this, at least in
firefox.

* chore: include a css reset

There has been some render inconsistency, especially on mobile, and I
suppose it has to do with the fact there was no CSS reset happening.

Just a guess. We'll see how things look with a reset in place and take
next steps as needed.